+++ build.ps1
@@ -22,7 +22,6 @@ pyinstaller --noconfirm $mode --windowed `
     --add-data "CHANGELOG.md;." `
     --add-data "CHANGELOG.fr.md;." `
     --add-data "USER_GUIDE.md;." `
-    --hidden-import soupsieve `
     --collect-all matplotlib `
     main.py